Move content in worker_config.py to scenarios.py
